pyproject.toml
[project]
name = "gpt-talker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "google-genai>=1.67.0",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
]
